servers: ds, rs, pm and sched build for aarch64
The first four boot-image servers now compile and link. Each stopped on
something different, and none of it was in the servers themselves except the
last two.

libminc was missing strcspn. On a 64-bit platform strcspn.c compiles to
nothing at all - it is guarded by "#if ULONG_MAX != 0xffffffffffffffffull"
and its own comment says the 64-bit version lives in strspn.c, which defines
both functions. libminc built strcspn.o, got an empty object, and ds failed
to link. strspn.o joins the list; on 32-bit ports it only adds strspn, which
nothing there was getting either.

alloca was declared, not defined. It cannot be a library function - it
allocates in its caller's frame - so the declaration in <stdlib.h> relies on
the compiler recognising the name and expanding the builtin. GCC 15 under
-fno-builtin, which is how libc and the servers are built, emits a call
instead, and rs failed to link on it through execle(). The builtin is now
named outright, outside the _NETBSD_SOURCE block: libminc compiles without
that macro and so never saw any of the declarations at all.

pm did not know the architecture: misc.c had #error for anything that was not
i386 or arm, in the table uname reports. It says evbarm64 - the MACHINE name,
the way the other two entries say evbarm and i386 rather than arm and x86.

pm's alarm clamp was written as LONG_MAX. clock_t is 32 bits on every port,
so on ILP32 that was the right number by coincidence; on LP64 it is 2^63-1,
and the clamp truncated to 0xffffffff - a larger value than the overflow it
guarded against, and negative when read back as a long. Written in terms of
clock_t it is 0x7fffffff everywhere, which is what the 32-bit ports already
got.

vfs needed a dozen small ones, all of the same two kinds. dev_t and ino_t are
64 bits wide on every port, but the 64-bit type is unsigned long long on
ILP32 and unsigned long on LP64, so ten %ll conversions were wrong; the
arguments are cast to what the format asks for. And several prototypes
disagreed with their definitions about array bounds, which GCC 15 now
compares - get_name, canonical_path and rdlink_direct gained the bounds their
definitions state, while mount_fs and map_driver lost bounds their callers
never honoured (main.c mounts the boot ramdisk with the literals
"bootramdisk" and "/").

One of those is a real defect rather than a diagnostic: FAILCHECK in
vfs/exec.c was written "if(...) { goto ...; } while(0)" - the "do" is
missing, so the macro was an if statement followed by an unrelated empty
loop. Only the first half would be governed by any enclosing if. It has
worked because every use stands alone as a statement.

vfs does not build yet; the next thing it stops on is coredump.c. mib is
untried.

Classification: [возраст] for alloca, the formats, the array bounds and
FAILCHECK; [порт] for strspn, the uname entry and the alarm clamp.

+++ minix/servers/vfs/filedes.c
@@ -507,8 +507,16 @@ close_filp(struct filp * f, int may_suspend)
 	f->filp_mode = FILP_CLOSED;
 	f->filp_count = 0;
   } else if (f->filp_count < 0) {
+	/*
+	 * dev_t and ino_t are 64 bits wide on every port, but the type that
+	 * is 64 bits is not the same one: unsigned long long on ILP32,
+	 * unsigned long on LP64. The %ll conversions here are right for the
+	 * first and wrong for the second, so the values are cast to what the
+	 * format actually asks for.
+	 */
 	panic("VFS: invalid filp count: %d ino %llx/%llu", f->filp_count,
-	      vp->v_dev, vp->v_inode_nr);
+	      (unsigned long long)vp->v_dev,
+	      (unsigned long long)vp->v_inode_nr);
   } else {
 	unlock_vnode(f->filp_vno);
   }
